Cast unsigned int

+++ WinScreen.cpp
@@ -24,11 +24,10 @@ void WinScreen::handle_input(SDL_Event e){
       if (m_name.length() == 0) {
           m_entered_invalid = true;
       } else {
-          m_score->add_score({m_name, m_timer->get_time()});
+          m_score->add_score({m_name, (unsigned int)m_timer->get_time()});
           m_score->save_scores();
           m_mode = game_modes::HIGHSCORES;
-      }
-    } else if(e.key.keysym.sym == SDLK_BACKSPACE){
+      }    } else if(e.key.keysym.sym == SDLK_BACKSPACE){
       if (m_name.size() > 0){
         m_name.pop_back();
       }
